add file: Build and deploy to GitHub Pages
