+++ .github/workflows/deploy.yml
@@ -1,0 +1,26 @@
+name: Build and deploy to GitHub Pages
+
+on:
+  push:
+    branches: [main]
+env:
+  VITE_API_TEST: ${{secrets.VITE_API_TEST}}
+
+jobs:
+  build-and-deploy:
+    runs-on: ubuntu-latest
+    steps:
+      - name: Checkout 🛎️
+        uses: actions/checkout@v3
+
+      - name: Install, build 🔧
+        run: |
+          npm install
+          npm run build
+          cp ./dist/index.html ./dist/404.html
+
+      - name: Deploy 🚀
+        uses: JamesIves/github-pages-deploy-action@4.1.0
+        with:
+          branch: gh-pages
+          folder: dist
